refactor : html 파일 및 렌더링 코드 수정
