+++ src/main/java/codesquad/business/configuration/handlerMapConfig.java
@@ -15,6 +15,7 @@ public class handlerMapConfig {
         handlerMap.setHandlerMap("/article/list", ArticleListHandler.articleListHandler);
         handlerMap.setHandlerMap("/article/detail", ArticleDetailHandler.articleDetailHandler);
         handlerMap.setHandlerMap("/article/comment", CommentHandler.commentHandler);
+        handlerMap.setHandlerMap("/logout", LogoutHandler.logoutHandler);
     }
 
     private handlerMapConfig() {}
