+++ src/main/java/codesquad/was/render/HtmlTemplateRender.java
@@ -2,8 +2,8 @@ package codesquad.was.render;
 
 public class HtmlTemplateRender {
     public static final String authButtonHolder = "{{authButton}}";
-    private static ListRender listRender = new ListRender();
-    private static SingleRender singleRender = new SingleRender();
+    private static final ListRender listRender = new ListRender();
+    private static final SingleRender singleRender = new SingleRender();
 
     public static String render(String html, Model model) {
         html = authButtonRender(html, model);
@@ -14,30 +14,31 @@ public class HtmlTemplateRender {
 
     private static String authButtonRender(String html, Model model) {
         String authButton = """
-          <li class="header__menu__item">
-            <a class="btn btn_contained btn_size_s" href="/login">로그인</a>
-          </li>
-          <li class="header__menu__item">
-            <a class="btn btn_ghost btn_size_s" href="/registration">
-              회원 가입
-            </a>
-          </li>
-        """;
+                  <li class="header__menu__item">
+                    <a class="btn btn_contained btn_size_s" href="/login">로그인</a>
+                  </li>
+                  <li class="header__menu__item">
+                    <a class="btn btn_ghost btn_size_s" href="/registration">
+                      회원 가입
+                    </a>
+                  </li>
+                """;
 
         String userName = (String) model.getSingleData().get("userName");
-        if(userName !=null){
+        if (userName != null) {
             authButton = """
-            <li class="header__menu__item">
-              <a class="btn btn_contained btn_size_s" href="/article">글쓰기</a>
-            </li>
-            <li class="header__menu__item">
-              <button class="btn btn_ghost btn_size_s">""" + userName + """
-            </li>
-            <li class="header__menu__item">
-              <button id="logout-btn" class="btn btn_ghost btn_size_s">로그아웃</button>
-            </button>
-            </li>
-            """;
+                            <li class="header__menu__item">
+                              <a class="btn btn_contained btn_size_s" href="/article">글쓰기</a>
+                            </li>
+                            <li class="header__menu__item">
+                              <button class="btn btn_ghost btn_size_s"> {{userName}}</button>
+                            </li>
+                            <li class="header__menu__item">
+                              <form action="/logout" method="POST" style="display: inline;">
+                                <button type="submit" id="logout-btn" class="btn btn_ghost btn_size_s">로그아웃</button>
+                              </form>
+                            </li>
+                    """;
         }
 
         return html.replace(authButtonHolder, authButton);
