+++ src/main/resources/static/articleDetails/index.html
@@ -6,6 +6,39 @@
   <link href="../reset.css" rel="stylesheet" />
   <link href="../global.css" rel="stylesheet" />
   <link href="../main.css" rel="stylesheet" />
+  <style>
+    .comments {
+      margin-top: 20px;
+    }
+    .comment {
+      padding: 10px 0;
+      border-bottom: 1px solid #ddd;
+    }
+    .comment p {
+      margin: 5px 0;
+    }
+    .comment-form {
+      margin-top: 20px;
+    }
+    .comment-form textarea {
+      width: 100%;
+      padding: 10px;
+      border: 1px solid #ddd;
+      border-radius: 4px;
+    }
+    .comment-form button {
+      margin-top: 10px;
+      background-color: #007BFF;
+      color: white;
+      padding: 10px 20px;
+      border: none;
+      border-radius: 4px;
+      cursor: pointer;
+    }
+    .comment-form button:hover {
+      background-color: #0056b3;
+    }
+  </style>
 </head>
 <body>
 <div class="container">
@@ -55,6 +88,14 @@
         </p>
       </li>
       $]
+      <div class="comment-form">
+        <h3>댓글 작성</h3>
+        <form action="/article/comment" method="post">
+          <textarea name="contents" placeholder="댓글을 입력하세요" required></textarea>
+          <input type="hidden" name="articleId" value="{{article.id}}">
+          <button type="submit">댓글 작성</button>
+        </form>
+      </div>
       <li class="comment__item hidden">
         <div class="comment__item__user">
           <img class="comment__item__user__img" />
@@ -76,35 +117,8 @@
         </div>
         <p class="comment__item__article">Comment 3</p>
       </li>
-      <button id="show-all-btn" class="btn btn_ghost btn_size_m">
-        모든 댓글 보기(3개)
-      </button>
     </ul>
-    <nav class="nav">
-      <ul class="nav__menu">
-        <li class="nav__menu__item">
-          <a class="nav__menu__item__btn" href="">
-            <img
-                    class="nav__menu__item__img"
-                    src="../img/ci_chevron-left.svg"
-            />
-            이전 글
-          </a>
-        </li>
-        <li class="nav__menu__item">
-          <a class="btn btn_ghost btn_size_m">댓글 작성</a>
-        </li>
-        <li class="nav__menu__item">
-          <a class="nav__menu__item__btn" href="">
-            다음 글
-            <img
-                    class="nav__menu__item__img"
-                    src="../img/ci_chevron-right.svg"
-            />
-          </a>
-        </li>
-      </ul>
-    </nav>
+
   </div>
 </div>
 </body>
